Extract parsing logic from tree rendering logic
